DateTime Functionality (RTC)
Added DateTime functionality as the Core2 has a RTC and v1.1 even has a separate RTC battery.

First add datetime to the ino before flashing to Core2. If the RTC battery runs out or the Core2 gets out of sync, add datetime to "datetime.txt" on SD card and select new menu option to update rtc datetime.

Additionally, when probe sniffing, when saving to SD card, probes along with the datetime they were sniffed are saved to "probeswdatetime.txt" on the SDcard.

Also some very minor spelling errors in comments.